Add i18n message and merge upload_table and recent_uploads_handler
