+++ app/assets/javascripts/components/activity/recent_uploads_handler.jsx
@@ -2,9 +2,41 @@ import React from 'react';
 import createReactClass from 'create-react-class';
 import PropTypes from 'prop-types';
 import { connect } from "react-redux";
-import UploadTable from './upload_table.jsx';
 import { fetchRecentUploads, sortRecentUploads } from '../../actions/recent_uploads_actions.js';
+import Loading from '../common/loading.jsx';
+import Upload from '../uploads/upload.jsx';
 
+
+export const UploadTable = createReactClass({
+  displayName: 'UploadTable',
+
+  propTypes: {
+    loading: PropTypes.bool,
+    uploads: PropTypes.array,
+  },
+
+  _renderUploads() {
+    return this.props.uploads.map((upload) => {
+      return (
+        <Upload upload={upload} key={upload.id} linkUsername={true} />
+      );
+    });
+  },
+
+  render() {
+    if (this.props.loading) {
+      return <Loading />;
+    }
+
+    const uploads = this._renderUploads();
+
+    return (
+      <div className="gallery">
+        {uploads}
+      </div>
+    );
+  }
+});
 
 export const RecentUploadsHandlerBase = createReactClass({
   displayName: 'RecentUploadsHandler',
