+++ app/assets/stylesheets/modules/_uploads.styl
@@ -43,7 +43,6 @@
 
 .info p.usage {
   font-size: 1.5em;
-  line-height: 3em;
   color: rgb(190, 151, 225);
 }
