+++ test/components/activity/recent_uploads_handler.spec.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import sinon from 'sinon';
-
+import ReactTestUtils from 'react-dom/test-utils';
 import '../../testHelper';
+import { UploadTable, RecentUploadsHandlerBase } from '../../../app/assets/javascripts/components/activity/recent_uploads_handler.jsx';
 
-import { RecentUploadsHandlerBase } from '../../../app/assets/javascripts/components/activity/recent_uploads_handler.jsx';
 
 describe('RecentUploadsHandler', () => {
   it('fetches recent uploads', () => {
@@ -14,5 +14,54 @@ describe('RecentUploadsHandler', () => {
 
     // called once when mounted
     expect(spy.callCount).to.eq(1);
+  });
+});
+
+
+describe('UploadTable', () => {
+  const uploads = [{
+    id: 123,
+    file_name: 'Paper prototype of website user interface, 2015-04-16.jpg',
+    thumburl: 'https://upload.wikimedia.org/wikipedia/commons/thumb/6/67/Paper_prototype_of_website_user_interface%2C_2015-04-16.jpg/640px-Paper_prototype_of_website_user_interface%2C_2015-04-16.jpg',
+    url: 'https://commons.wikimedia.org/wiki/File:Paper_prototype_of_website_user_interface,_2015-04-16.jpg',
+    uploader: 'Ragesoss',
+    uploaded_at: new Date().toISOString(),
+    usage_count: 1
+  }, {
+    id: 456,
+    file_name: 'Young man in bar, cross processed.JPG',
+    thumburl: 'https://upload.wikimedia.org/wikipedia/commons/thumb/c/c6/Young_man_in_bar%2C_cross_processed.JPG/640px-Young_man_in_bar%2C_cross_processed.JPG',
+    url: 'https://commons.wikimedia.org/wiki/File:Young_man_in_bar,_cross_processed.JPG',
+    uploader: 'Ragesoss',
+    uploaded_at: new Date().toISOString(),
+    usage_count: 2
+  }];
+
+
+  it('shows loading when loading attribute is true', () => {
+    const TestTable = ReactTestUtils.renderIntoDocument(
+      <div>
+        <UploadTable
+          loading={true}
+        />
+      </div>
+    );
+
+    const loading = TestTable.querySelector('.loading');
+    expect(loading).to.exist;
+  });
+
+  it('renders activities', () => {
+    const TestTable = ReactTestUtils.renderIntoDocument(
+      <div>
+        <UploadTable
+          loading={false}
+          uploads={uploads}
+        />
+      </div>
+    );
+
+    const bodyElement = TestTable.querySelectorAll('div.upload');
+    expect(bodyElement).to.exist;
   });
 });
